implement canonical_email user attribute
`canonical_email` is the "From:" address that is supposed to be used when users send email.

Historically it has been a requirement from leadership that IceCube's mail server use canonical addresses of the form firstname.lastname@icecube.wisc.edu, rather than username@icecube.wisc.edu.

`canonical_email` attribute records what this canonical address is for a user. Note that the algorithm used to generate the address from the user's name hasn't remained constant, and some users have completely custom canonical addresses, so the only way to be sure what the canonical address should be is to use this attribute.